requirements.txt
aiogram
sqlalchemy
asyncpg
envserv
notion-database
pillow